requirements.txt
anyio==4.1.0
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
emoji-country-flag==1.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
kink==0.7.0
python-dotenv==1.0.0
python-i18n==0.3.9
python-telegram-bot==20.7
requests==2.32.3
sniffio==1.3.0
standardjson==0.3.1
tornado==6.3.3
typing_extensions==4.8.0
urllib3==2.1.0
webhooks==0.4.2
wrapt==1.16.0
